+++ lib/rubocop/cop/lint/end_alignment.rb
@@ -5,8 +5,8 @@ module RuboCop
     module Lint
       # This cop checks whether the end keywords are aligned properly.
       #
-      # Three modes are supported through the `AlignWith` configuration
-      # parameter:
+      # Three modes are supported through the `EnforcedStyleAlignWith`
+      # configuration parameter:
       #
       # If it's set to `keyword` (which is the default), the `end`
       # shall be aligned with the start of the keyword (if, class, etc.).